+++ src/components/Notification/index.js
@@ -1,0 +1,7 @@
+export {
+  ToastNotification,
+  InlineNotification,
+  NotificationActionButton,
+  NotificationButton,
+  NotificationTextDetails,
+} from 'carbon-components-react';
