+++ src/components/ProgressIndicator/ProgressIndicator.story.jsx
@@ -4,9 +4,9 @@ import { boolean, number, select } from '@storybook/addon-knobs';
 import { storiesOf } from '@storybook/react';
 import { action } from '@storybook/addon-actions';
 
-import { ProgressIndicatorSkeleton } from '../..';
-
 import ProgressIndicator from './ProgressIndicator';
+
+import { ProgressIndicatorSkeleton } from '.';
 
 const items = [
   {
