+++ src/components/ProgressIndicator/index.js
@@ -1,0 +1,1 @@
+export { ProgressStep, ProgressIndicatorSkeleton } from 'carbon-components-react';
